backend/requirements.txt
# Core Django
Django==5.0.3
gunicorn==21.2.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
dj-database-url==2.3.0

# DRF and Auth
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0  # Remove rest-framework-simplejwt

# Database/Caching
django-redis==5.4.0
redis==6.2.0

# Dev/Utils (optional)
django-debug-toolbar==5.2.0
django-extensions==4.1
python-dotenv==1.1.0
django-cors-headers==4.7.0

# Fix potential troublemakers
protobuf==3.20.3  # Downgraded
numpy==2.0.0  # More stable version

python-decouple==3.8
drf-spectacular==0.28.0
django-filter==25.1
djangorestframework-gis==1.2.0
whitenoise==6.9.0
pillow==11.2.1
cloudinary==1.44.1
django-cloudinary-storage==0.3.0